replace coinOutputs and itemOutputs to entries

+++ x/pylons/handlers/execute_recipe.go
@@ -13,6 +13,32 @@ import (
 type ExecuteRecipeResp struct {
 	Message string
 	Status  string
+	Output  types.WeightedParam
+}
+
+func AddExecutedResult(ctx sdk.Context, keeper keep.Keeper, output types.WeightedParam, sender sdk.AccAddress, cbID string) sdk.Error {
+	switch output.(type) {
+	case types.CoinOutput:
+		if coinOutput, ok := output.(types.CoinOutput); ok {
+			var ocl sdk.Coins
+			ocl = append(ocl, sdk.NewCoin(coinOutput.Coin, sdk.NewInt(coinOutput.Count)))
+
+			_, _, err := keeper.CoinKeeper.AddCoins(ctx, sender, ocl)
+			if err != nil {
+				return err
+			}
+		}
+	case types.ItemOutput:
+		if itemOutput, ok := output.(types.ItemOutput); ok {
+			outputItem := *itemOutput.Item(cbID, sender)
+			if err := keeper.SetItem(ctx, outputItem); err != nil {
+				return sdk.ErrInternal(err.Error())
+			}
+		}
+	default:
+		return sdk.ErrInternal("no item nor coin type created")
+	}
+	return nil
 }
 
 // HandlerMsgExecuteRecipe is used to create cookbook by a developer
@@ -34,36 +60,11 @@ func HandlerMsgExecuteRecipe(ctx sdk.Context, keeper keep.Keeper, msg msgs.MsgEx
 	}
 
 	// output coins to the sender
-	var ocl sdk.Coins
 
-	for _, out := range recipe.CoinOutputs {
-		ocl = append(ocl, sdk.NewCoin(out.Coin, sdk.NewInt(out.Count)))
-	}
-
-	// store the execution as the interval
-	if recipe.BlockInterval > 0 {
-		exec.RecipeID = recipe.ID
-		exec.CoinInputs = cl
-		exec.CoinOutputs = ocl
-		exec.Sender = msg.Sender
-		exec.BlockHeight = ctx.BlockHeight() + recipe.BlockInterval
-	} else {
-
-		if !keeper.CoinKeeper.HasCoins(ctx, msg.Sender, cl) {
-			return sdk.ErrInternal("insufficient coin balance").Result()
-		}
-		// TODO: send the coins to a master address instead of burning them
-		// think about making this adding and subtracting atomic using inputoutputcoins method
-		_, _, err = keeper.CoinKeeper.SubtractCoins(ctx, msg.Sender, cl)
-		if err != nil {
-			return err.Result()
-		}
-		_, _, err = keeper.CoinKeeper.AddCoins(ctx, msg.Sender, ocl)
-		if err != nil {
-			return err.Result()
-		}
-
-	}
+	// TODO: CoinOutputs and ItemOutputs should be merged so just need to update this
+	// for _, out := range recipe.CoinOutputs {
+	// 	ocl = append(ocl, sdk.NewCoin(out.Coin, sdk.NewInt(out.Count)))
+	// }
 
 	// Item transaction
 	// first lets handle all item to item transactions
@@ -107,52 +108,64 @@ func HandlerMsgExecuteRecipe(ctx sdk.Context, keeper keep.Keeper, msg msgs.MsgEx
 	}
 
 	// TODO: validate 1-1 correspondence for item input and output - check ids
-	var outputItems []types.Item
-	for _, item := range recipe.ItemOutputs {
-		outputItems = append(outputItems, *item.Item(recipe.CookbookId, msg.Sender))
-	}
 
 	// we set the inputs and outputs for storing the execution
 	if recipe.BlockInterval > 0 {
+		// store the execution as the interval
+		exec.RecipeID = recipe.ID
+		exec.CookbookID = recipe.CookbookId
+		exec.CoinInputs = cl
+		exec.Sender = msg.Sender
+		exec.BlockHeight = ctx.BlockHeight() + recipe.BlockInterval
 		exec.ItemInputs = matchedItems
-		exec.ItemOutputs = outputItems
+		exec.Entries = recipe.Entries
 		exec.ID = exec.KeyGen()
 		err2 := keeper.SetExecution(ctx, exec)
 
 		if err2 != nil {
 			return sdk.ErrInternal(err2.Error()).Result()
 		}
-		resp, err2 := json.Marshal(ExecuteRecipeResp{
+		resp, err3 := json.Marshal(ExecuteRecipeResp{
 			Message: "scheduled the recipe",
 			Status:  "Success",
 		})
 
+		if err3 != nil {
+			return sdk.ErrInternal(err2.Error()).Result()
+		}
+		return sdk.Result{Data: resp}
+	} else {
+		if !keeper.CoinKeeper.HasCoins(ctx, msg.Sender, cl) {
+			return sdk.ErrInternal("insufficient coin balance").Result()
+		}
+		// TODO: send the coins to a master address instead of burning them
+		// think about making this adding and subtracting atomic using inputoutputcoins method
+		_, _, err = keeper.CoinKeeper.SubtractCoins(ctx, msg.Sender, cl)
+		if err != nil {
+			return err.Result()
+		}
+
+		// we delete all the matched items as those get converted to output items
+		for _, item := range matchedItems {
+			keeper.DeleteItem(ctx, item.ID)
+		}
+
+		output := recipe.Entries.Actualize()
+		err = AddExecutedResult(ctx, keeper, output, msg.Sender, recipe.CookbookId)
+
+		if err != nil {
+			return err.Result()
+		}
+
+		resp, err2 := json.Marshal(ExecuteRecipeResp{
+			Message: "successfully executed the recipe",
+			Status:  "Success",
+			Output:  output,
+		})
+
 		if err2 != nil {
 			return sdk.ErrInternal(err2.Error()).Result()
-
 		}
 		return sdk.Result{Data: resp}
 	}
-
-	// we delete all the matched items as those get converted to output items
-	for _, item := range matchedItems {
-		keeper.DeleteItem(ctx, item.ID)
-	}
-
-	for _, item := range outputItems {
-		if err := keeper.SetItem(ctx, item); err != nil {
-			return sdk.ErrInternal(err.Error()).Result()
-		}
-	}
-
-	resp, err2 := json.Marshal(ExecuteRecipeResp{
-		Message: "successfully executed the recipe",
-		Status:  "Success",
-	})
-
-	if err2 != nil {
-		return sdk.ErrInternal(err2.Error()).Result()
-
-	}
-	return sdk.Result{Data: resp}
 }
